ExeLounge/groupisite/requirements.txt
asgiref
asgi_redis
astroid
better-profanity
bleach
channels
channels_redis
colorama
dj-static
Django
django-material
django-recaptcha
django_heroku
gunicorn
isort
lazy-object-proxy
mccabe
Pillow
psycopg2-binary
pylint
pytz
six
sqlparse
static3
toml
wrapt
